Implemented Mix-Down Frequency
With this commit we add a Mix-Down Frequency of 61.74 MHz to mix down
the signal.
The mix frequency will manifest in an (omega1 - omega2) and
an (omega1 + omega2) frequenies. Because the electronics implements a
low band bandpass filter we average out the (omega1 + omega2)
oscillations.
We can do that on the individual cell level, because the flux in the
pickup coil, the scalar product of B*mu are all linear equations and we
can pull the sin(omega2) term into the function of each cell. There we
use the addition theorem for sin functions and get the
sin((omega1-omega2)*t) term and the sin((omega1+omega2)*t) term.
The later one is fast oscillating and averaged over several oscillation
periods the term vanishes. Thus we only have to modify the sin(omega1*t)
in the magnetization of the cells to a sin((omega1-omega2)*t).

We extend the plot of the FID over a longer time range. We get out a
perfect decaying FID as we put in a perfect homogenous external field.

Minor clean up in the commented out part of the code.