Improved performance of FID calculation
With this commit we implemented a more efficient way to calculate the
FID signal. We convert the cell dependent quantities in numpy arrays of
dimention `NCell`. The time variable t can now be given as 1D numpy
array instead of scalars. The time variable has dimension `NTime`.
The final formula is reformulated so that we have less calls to
expensive function calls, such as exp, sin, arctan, ... .
Parts of the formula that are used several times are precomputed.
The result of the function is now a 1D numpy array of dimension `NTime`
instead of a scalar.
The output FID signal is unchanged.
The calculation time reduces from > 700 sec for 10000 points to
1.4 seconds, which is a spead up of about 500. The FID calculation is
no longer the most time consuming part.
--> Implement the cell parameters directly as numpy arrays instead of
a list with cell properties.
--> See if other parts of the code can be accelerated.