Added noise classe
With this commit we add a Noise class, which allows to generate Noise
with different frequency powers in spectral power density.
The scale of the noise has to be matched vs the flux.
We add the noise to the plot flux vs time. Effects can be seen in power
spectrum. Adding noise currently not performance relevant.

The FFT is pulled out of the plotting part because we always want to do
the FFT of the signal.

The digitization rate is 10 MSPS for fixed probes (which is heavily
oversampled) and 1 MSPS for trolly. Here we use currently 1 MSPS. We are
not memory limited at the moment.