Implemented pickup_flux in a way that we do not exceed the memory
We evaluate only time chuncks of the full time array at a time, so
that we do not exceed the available memory. The chunck size is defined
by available memory. By testing this we noticed that the pickup_flux
depends on the number of cells, because the sum over the cells just
increases and magnetization does not scales with per cell. We have to
use the sum in pickup_flux like an average and devid by the number of
cells. This brings the pickup_flux to muV which is also resonable.
However we have to think about that a bit better.