Seperated Class implementation and analysis script in seperate files.
With this commit we seperate the class definition which will live in 
FID_Simulation and the analysis script which live in analysis.py. 
Additionally we collect FFT related functions in FFT_analysis.py. We 
also set that the t_90 is calculated by default if an RF_pulse is 
applied if no time is specified.