Implemented self contribution from magnetization to b-field
With this commit:
* we add the ppm, ppb, ppt, ... and T0 as an extension to the 
numericalunits package, so that we can use this notation in the code. 
Here T0 = -273.15K is the absolute zero of the Celsius scale.
* We changed the code in several places to make use of the new definde 
pseudo units
* we made the time_step of the RK solver an argument of the 
apply_rf_field_nummerical function, so that we can modify it at runtime
* we added an argument to apply_rf_field_nummerical called 
with_self_contribution, which is a flag that indicates if we should 
consider the additional field from the magnetization of the cell. We 
estimate that this is a 2 ppb effect.