Seperated classes into individual files for better structure of package
With this commit we seperate the code into different files to better 
structure the code.

Unit definitions of into unit and we import only units defined in unit 
here. This allows us to factorize the units better if we want to change 
the way we handle units.

Materials are defined in material.py. This contains the main class as 
well as some pre defined materials.

Noise.py contains classes to generate noise for time traces

Magnet.py contains classes related to the magnet.

FID_simulation.py contains definition of Coil and Probe

E989.py contains specialliezed classes for the Muon g-2 E989 project at 
Fermilab