Fixed FID generation with multiple chuncks.
Problem: if multiple chuncks had to be used due to finite memory, the 
time of the next chunck was shifted by one time step in the time series 
because we made the reference time the first time of the chunck instead 
of the last time of the previous chunck. That resulted in steps in the 
FID and thus in the envelope and phase function.
Solution: We introduce a t0 variable to keep track of the t0 time of the 
phase0 argument. 
Test: Tested in jupyter notebook with different NCell values (chuncks 
from 1 to ~10) and looking at difference between these simulations. No 
discontinuities visible. Only modulation due to increased cells visible.