Fixed baseline extraction by handeling special edge cases.
* With this commit we fix the baseline extraction of an FID
* remove return_derivative argument from PhaseFunction, now we save 
these arrays as attributes of the class and do not need to change the 
return type.
* saveral arrays are now attributes of the class to excess them for 
debugging or analysis
* note that np.sign can result in -1, 0, 1 and not only in -1, 1.
* added a seperate check if sign is 0. In this case we are at a zero 
crossing and do not have to interpolate. Slope is taken from point 
before and after the current point.
* derivatives can be zero and thus 1/derivative leads to inf and nan 
values. We now ignore them in the construction of the spline because 
otherwise the spline only returns nan.
* Added that the spline should return 0 in case of extrapolation
* Tested this implementation on several trolley FID and it seems it is 
robust and does not throw errors. However it only makes sens if the 
signal is not completely below zero.